pyproject.toml
[tool.poetry]
name = "calculator"
version = "0.1.0"
description = ""
authors = ["iscanderoff <scanderoff@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.79.1"
uvicorn = "^0.18.2"
psycopg2 = "^2.9.3"
SQLAlchemy = "^1.4.40"
email-validator = "^1.2.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.5"
pydantic = {extras = ["dotenv"], version = "^1.9.2"}
alembic = "^1.8.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
